First take on refactoring
- simplified messaging between content script and the extension
- using classes and ES6 modules
- started breaking down code into logical pieces
- use lightweight zora+esm combo instead of mocha for testing
