+++ .eslintrc.json
@@ -1,7 +1,8 @@
 {
   "extends": "eslint:recommended",
   "parserOptions": {
-    "ecmaVersion": 6
+    "ecmaVersion": 6,
+    "sourceType": "module"
   },
   "globals": {
     "di": true,
