--- src/background.html
@@ -1,1 +1,0 @@
-<script src="background.js"></script>
