+++ src/background.js
@@ -1,11 +1,11 @@
 chrome.runtime.onMessage.addListener(
   (arg) => {
-    if (arg.type && (arg.type === "LeoExportExtension.DownloadCSV")) {
-      var resultBlob = new Blob(arg.payload, { type: "text/plain;charset=utf-8" })
-      var url = URL.createObjectURL(resultBlob);
+    if (arg.type && (arg.type === "ankileo.download")) {
+      const resultBlob = new Blob(arg.payload.data, { type: "text/plain;charset=utf-8" })
+      const url = URL.createObjectURL(resultBlob);
       chrome.downloads.download({
         url: url,
-        filename: 'lingualeo-dict-export.csv'
+        filename: arg.payload.name,
       });
     }
   }
