+++ src/js/all.test.js
@@ -1,0 +1,3 @@
+require = require("esm")(module); // eslint-disable-line no-undef
+require('./pageMatcher.test.js');
+require('./util.test.js');
