+++ src/js/button.html.js
@@ -1,0 +1,22 @@
+const html = `
+<a class="btn leo-export-extension-btn" id="leo-export-extension-btn">{Export}</a>
+<ul class="leo-export-extension-menu-container" style="display:none">
+  <li class="active">
+    <a href="javascript: void 0" id="leo-export-extension-btn-all">
+      <i class="iconm-none"></i>{All}
+    </a>
+  </li>
+  <li class="active">
+    <a href="javascript: void 0" id="leo-export-extension-btn-new">
+      <i class="iconm-w-big-25"></i>{Learning}
+    </a>
+  </li>
+  <li class="active">
+    <a href="javascript: void 0" id="leo-export-extension-btn-selected">
+      <i class="iconm-checklight"></i>{Selected}
+    </a>
+  </li>
+</ul>
+`;
+
+export default html;
