+++ src/js/locale.js
@@ -1,0 +1,13 @@
+import { format } from './util';
+
+export default class Locale {
+	constructor(lang = 'en', translations) {
+    this.lang = lang;	
+    this.translations = translations;
+  }
+
+	t(key, params) {
+		const str = this.translations[this.lang].translation[key];
+		return format(str, params)
+	}
+}
