+++ src/js/pageMatcher.test.js
@@ -1,12 +1,11 @@
-window = {}; // eslint-disable-line no-global-assign
-require('./PageWithWordsMatcher.js');
-var matcher = window.LeoExport.PageWithWordsMatcher;
+require = require("esm")(module); // eslint-disable-line no-undef
+const describe = require("zora").test;
+const PageMatcher = require("./pageMatcher").default;
 
-var assert = require('assert');
+var matcher = new PageMatcher();
 
-describe('PageWithWordsMatcher', () => {
-
-  describe('getWordGroupId', () => {
+describe('PageMatcher', t => {
+  t.test('getWordGroupId', t => {
     var tests = [
       { url: '/ru/userdict',                  expected: '' },
       { url: '/ru/userdict/wordSets/6505',    expected: '6505' },
@@ -19,11 +18,12 @@ describe('PageWithWordsMatcher', () => {
       { url: '/ru/glossary/wordSets/6505',    expected: '6505' },
       { url: '/ru/glossary/learn/6506',       expected: '6506' },
     ];
-
-    tests.forEach((test) => {
-      it('should ' + (test.expected === false ? 'NOT ' : '') + 'match ' + test.url, () => {
-        assert.equal(matcher.getWordGroupId(test.url), test.expected);
+  
+    tests.forEach(fxt => {
+      t.test('should ' + (fxt.expected === false ? 'NOT ' : '') + 'match ' + fxt.url, t => {
+        t.equal(matcher.getWordGroupId(fxt.url), fxt.expected);
       });
     });
-  });
-})
+  });  
+});
+
