+++ src/js/util.test.js
@@ -1,0 +1,13 @@
+require = require("esm")(module); // eslint-disable-line no-undef
+const describe = require("zora").test;
+const util = require("./util");
+
+describe('Util', t => {
+  t.test('format', t => {
+    const { format } = util;
+    t.equal(format("Test"), "Test");
+    t.equal(format("Test {one}", { one: 1 }), "Test 1");
+    t.equal(format("Test {one} and {two}", { one: 1, two: 2 }), "Test 1 and 2");
+    t.equal(format("Test {one} and {two}", { one: 1 }), "Test 1 and {two}");
+  });
+});
